pyproject.toml
[project]
name = "rfb"
version = "0.1.0"
description = ""
authors = [
    {name = "arthraw", email = "arthursallesandrade@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "scrapy (>=2.14.1,<3.0.0)",
    "apache-airflow (>=3.1.7,<4.0.0)",
    "azure-storage-file-datalake (>=12.23.0,<13.0.0)",
    "azure-identity (>=1.25.2,<2.0.0)",
    "dbt-core (>=1.11.6,<2.0.0)",
    "databricks-sdk (>=0.93.0,<0.94.0)",
    "databricks-cli (>=0.18.0,<0.19.0)",
]

[tool.poetry]
packages = [{include = "rfb", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
